requirements.txt
cx-Oracle>=7.0.0
jxa-jkedra==0.0.1
